+++ core/src/muxing/boxed.rs
@@ -6,10 +6,11 @@ use std::{
     task::{Context, Poll},
 };
 
+use bytes::Bytes;
 use futures::{AsyncRead, AsyncWrite};
 use pin_project::pin_project;
 
-use crate::muxing::{StreamMuxer, StreamMuxerEvent};
+use crate::muxing::{SendDatagramError, StreamMuxer, StreamMuxerEvent};
 
 /// Abstract `StreamMuxer`.
 pub struct StreamMuxerBox {
@@ -26,7 +27,7 @@ impl fmt::Debug for StreamMuxerBox {
 ///
 /// A [`SubstreamBox`] erases the concrete type it is given and only retains its `AsyncRead`
 /// and `AsyncWrite` capabilities.
-pub struct SubstreamBox(Pin<Box<dyn AsyncReadWrite + Send>>);
+pub struct SubstreamBox(Pin<Box<dyn AsyncReadWrite + Send>>, Option<u64>);
 
 #[pin_project]
 struct Wrap<T>
@@ -53,7 +54,7 @@ where
         self.project()
             .inner
             .poll_inbound(cx)
-            .map_ok(SubstreamBox::new)
+            .map_ok(|s| SubstreamBox::new(T::substream_id(&s), s))
             .map_err(into_io_error)
     }
 
@@ -64,7 +65,7 @@ where
         self.project()
             .inner
             .poll_outbound(cx)
-            .map_ok(SubstreamBox::new)
+            .map_ok(|s| SubstreamBox::new(T::substream_id(&s), s))
             .map_err(into_io_error)
     }
 
@@ -78,6 +79,14 @@ where
         cx: &mut Context<'_>,
     ) -> Poll<Result<StreamMuxerEvent, Self::Error>> {
         self.project().inner.poll(cx).map_err(into_io_error)
+    }
+
+    fn send_datagram(self: Pin<&mut Self>, data: Bytes) -> Result<(), SendDatagramError> {
+        self.project().inner.send_datagram(data)
+    }
+
+    fn max_datagram_size(&self) -> Option<usize> {
+        self.inner.max_datagram_size()
     }
 }
 
@@ -139,13 +148,30 @@ impl StreamMuxer for StreamMuxerBox {
     ) -> Poll<Result<StreamMuxerEvent, Self::Error>> {
         self.project().poll(cx)
     }
+
+    fn send_datagram(self: Pin<&mut Self>, data: Bytes) -> Result<(), SendDatagramError> {
+        self.project().send_datagram(data)
+    }
+
+    fn max_datagram_size(&self) -> Option<usize> {
+        self.inner.as_ref().get_ref().max_datagram_size()
+    }
+
+    fn substream_id(substream: &Self::Substream) -> Option<u64> {
+        substream.transport_stream_id()
+    }
 }
 
 impl SubstreamBox {
     /// Construct a new [`SubstreamBox`] from something
     /// that implements [`AsyncRead`] and [`AsyncWrite`].
-    pub fn new<S: AsyncRead + AsyncWrite + Send + 'static>(stream: S) -> Self {
-        Self(Box::pin(stream))
+    pub fn new<S: AsyncRead + AsyncWrite + Send + 'static>(id: Option<u64>, stream: S) -> Self {
+        Self(Box::pin(stream), id)
+    }
+
+    /// Transport-assigned stream id, see [`StreamMuxer::substream_id`].
+    pub fn transport_stream_id(&self) -> Option<u64> {
+        self.1
     }
 }